requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6
groq==0.22.0
pypdf==3.17.1
python-docx==1.0.1
pandas==2.1.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4 